requirements.txt
fastapi==0.115.5
google_cloud_aiplatform==1.73.0
google-cloud-pubsub==2.27.1
google-cloud-storage==2.18.2
numpy==1.26.4
protobuf==4.21.6
pydantic==2.9.2
python-dotenv==1.0.1
tensorflow==2.17.1
uvicorn==0.15.0
PySastrawi==1.2.0
joblib==1.4.2
